efafads

+++ longfirstgenmsg.py
@@ -60,8 +60,8 @@ def eltdt_rls(lat,long):
 
 def encodelongFGB(hex_code,latitude,southnorth,longitude,eastwest, suppdata):
 
-    latitude=float(latitude) * 1000
-    longitude = float(longitude) * 1000
+    latitude=int(float(latitude) * 1000)
+    longitude = int(float(longitude) * 1000)
     print(type(latitude))
     print(latitude)
     c = decodehex2.BeaconFGB()
